refactor: folder structure and file locations
