+++ src/App.tsx
@@ -3,9 +3,9 @@ import { MouseEventHandler, useCallback, useState } from "react";
 import i18n from "i18next";
 import { initReactI18next, useTranslation } from "react-i18next";
 import "./App.css";
-import translations from "./translations";
-import Form from "./Form";
-import Departures from "./Departures";
+import translations from "./libs/translations";
+import Form from "./components/Form";
+import Departures from "./components/Departures";
 import type { Search } from "./types";
 import Box from "@mui/material/Box";
 import AppBar from "@mui/material/AppBar";
